dev-llm/pyproject.toml
[project]
name = "vmaf-dev-llm"
version = "0.1.0"
description = "Local LLM dev helpers for the Lusoris VMAF fork (Ollama-first)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "BSD-3-Clause-Plus-Patent" }
authors = [
  { name = "Lusoris", email = "lusoris@pm.me" },
  { name = "Claude (Anthropic)", email = "noreply@anthropic.com" },
]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.6", "mypy>=1.10"]
cloud = [
  "anthropic>=0.40",
  "openai>=1.50",
]

[project.scripts]
vmaf-dev-llm = "vmaf_dev_llm.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vmaf_dev_llm"]

[tool.hatch.build.targets.wheel.force-include]
"src/vmaf_dev_llm/prompts" = "vmaf_dev_llm/prompts"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
